--- src/app/admin/admin.module.ts
@@ -1,40 +1,0 @@
-import { NgModule } from "@angular/core";
-import { CommonModule } from "@angular/common";
-import { RouterModule } from "@angular/router";
-import { AdminLayoutComponent } from './shared/components/admin-layout/admin-layout.component';
-import { LoginComponent } from './login/login.component';
-import { CreatePageComponent } from './create-page/create-page.component';
-import { FeedPageComponent } from './feed-page/feed-page.component';
-import { EditPageComponent } from './edit-page/edit-page.component';
-import { SharedModule} from "../shared/shared.module";
-
-@NgModule({
-  declarations: [
-    AdminLayoutComponent,
-    LoginComponent,
-    CreatePageComponent,
-    FeedPageComponent,
-    EditPageComponent
-  ],
-  imports: [
-    CommonModule,
-    RouterModule.forChild([
-      {
-        path: '', component: AdminLayoutComponent, children: [
-          {path: '', redirectTo: '/admin/login', pathMatch: 'full'},
-          {path: 'login', component: LoginComponent},
-          {path: 'feed', component: FeedPageComponent},
-          {path: 'create', component: CreatePageComponent},
-          {path: 'post/:id/edit', component: EditPageComponent},
-        ]
-      }
-    ]),
-    SharedModule
-  ],
-  exports: [RouterModule]
-})
-export class AdminModule {
-
-}
-
-
