--- src/app/admin/edit-page/edit-page.component.html
@@ -1,1 +1,0 @@
-<p>edit-page works!</p>
