--- src/app/admin/feed-page/feed-page.component.html
@@ -1,8 +1,0 @@
-<div
-  fxLayout="column"
-  fxLayoutAlign="center center"
-  fxLayoutGap="20px"
-  class="posts"
-  *ngFor="let post of posts">
-  <app-post [post]="post" [openButtonIsVisible]="false"></app-post>
-</div>
