--- src/app/admin/feed-page/feed-page.component.scss
@@ -1,3 +1,0 @@
-.posts {
-  margin: 10px;
-}
