--- src/app/admin/shared/components/admin-layout/admin-layout.component.html
@@ -1,22 +1,0 @@
-<mat-toolbar>
-  <div>Admin Blog</div>
-  <div fxFlex fxLayout fxLayoutAlign="flex-end">
-    <ul fxLayout fxLayoutGap="15px" class="navigation-items">
-      <li><a routerLinkActive="active" [routerLink]="['/admin', 'feed']">Read my posts</a></li>
-      <li><a routerLinkActive="active" [routerLink]="['/admin', 'create']">Create Post</a></li>
-      <li>
-        <a
-          routerLinkActive="active"
-          [routerLinkActiveOptions]={exact:true}
-          [routerLink]="['/']">All posts</a>
-      </li>
-      <li>
-        <a
-          routerLinkActive="active"
-          [routerLink]="['#']"
-          (click)="logout($event)">Logout</a>
-      </li>
-    </ul>
-  </div>
-</mat-toolbar>
-<router-outlet></router-outlet>
