--- src/app/admin/shared/components/admin-layout/admin-layout.component.scss
@@ -1,20 +1,0 @@
-mat-toolbar {
-  position: sticky;
-  top: 0;
-  z-index: 1000;
-}
-a {
-  text-decoration: none;
-  color: black;
-}
-
-.navigation-items {
-  list-style: none;
-  padding: 0;
-  margin: 0;
-}
-
-.active {
-  opacity: 0.5;
-}
-
