+++ src/app/app-routing.module.ts
@@ -1,20 +1,16 @@
 import { NgModule } from '@angular/core';
 import { PreloadAllModules, RouterModule, Routes} from '@angular/router';
-import { MainLayoutComponent } from "./shared/components/main-layout/main-layout.component";
-import { HomeComponent } from "./home/home.component";
-import { PostPageComponent } from "./post-page/post-page.component"
+import { PostsComponent } from "./posts/posts.component";
+import { PostPageComponent } from "./posts/post-page/post-page.component"
+import { LoginComponent } from "./auth/login/login.component";
+import { CreatePageComponent } from "./posts/create-page/create-page.component";
 
 const routes: Routes = [
-  {
-    path: '', component: MainLayoutComponent, children: [
-      {path: '', redirectTo: '/', pathMatch: 'full'},
-      {path: '', component: HomeComponent},
-      {path: 'post/:id', component: PostPageComponent }
-    ]
-  },
-  {
-    path: 'admin', loadChildren: ()=> import('./admin/admin.module').then(m=> m.AdminModule)
-  }
+    {path: '', redirectTo: '/', pathMatch: 'full'},
+    {path: '', component: PostsComponent},
+    {path: 'post/:id', component: PostPageComponent },
+    {path: 'login', component: LoginComponent},
+    {path: 'create', component: CreatePageComponent}
 ];
 
 @NgModule({
