+++ src/app/app.component.html
@@ -1,2 +1,6 @@
-<router-outlet></router-outlet>
+<app-header></app-header>
+<div class="main">
+    <router-outlet></router-outlet>
+</div>
+
