+++ src/app/app.component.scss
@@ -1,0 +1,4 @@
+.main {
+  width: 80%;
+  margin: 0 auto;
+}
