+++ src/app/app.module.ts
@@ -3,31 +3,51 @@ import { BrowserModule } from '@angular/platform-browser';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { HeaderComponent } from './shared/components/main-layout/header/header.component';
-import { MainLayoutComponent } from './shared/components/main-layout/main-layout.component';
-import { HomeComponent } from './home/home.component';
-import { PostPageComponent } from './post-page/post-page.component';
-import { SharedModule } from "./shared/shared.module";
+import { HeaderComponent } from './header/header.component';
+import { PostsComponent } from './posts/posts.component';
+import { PostPageComponent } from './posts/post-page/post-page.component';
+import { PostService } from "./posts/post.service";
+import { PostComponent } from "./posts/post/post.component";
+import { FlexLayoutModule } from "@angular/flex-layout";
+import { MatToolbarModule } from "@angular/material/toolbar";
+import { MatIconModule } from "@angular/material/icon";
+import { MatCardModule } from "@angular/material/card";
+import { MatButtonModule } from "@angular/material/button";
+import { FormsModule, ReactiveFormsModule } from "@angular/forms";
+import { MatFormFieldModule } from "@angular/material/form-field";
+import { MatInputModule } from "@angular/material/input";
+import { CreatePageComponent } from "./posts/create-page/create-page.component";
+import { LoginComponent } from "./auth/login/login.component";
 
 @NgModule({
   declarations: [
     AppComponent,
     HeaderComponent,
-    MainLayoutComponent,
-    HomeComponent,
+    PostsComponent,
     PostPageComponent,
+    PostComponent,
+    CreatePageComponent,
+    LoginComponent
 
   ],
   imports: [
     BrowserModule,
     BrowserAnimationsModule,
-    SharedModule,
-    AppRoutingModule
+    AppRoutingModule,
+    FlexLayoutModule,
+    MatToolbarModule,
+    MatIconModule,
+    MatCardModule,
+    MatButtonModule,
+    FormsModule,
+    ReactiveFormsModule,
+    MatFormFieldModule,
+    MatInputModule
   ],
   exports: [
     AppComponent,
-
   ],
-  bootstrap: [AppComponent]
+  providers: [ PostService ],
+  bootstrap: [ AppComponent ]
 })
 export class AppModule { }
