+++ src/app/auth/login/login.component.html
@@ -1,0 +1,39 @@
+<form
+  fxLayout="column"
+  fxLayoutAlign="center center"
+  fxLayoutGap="10px"
+  [formGroup]="loginForm"
+  (ngSubmit)="submit()">
+  <h2>Login into Admin Panel</h2>
+  <mat-form-field appearance="fill">
+    <input
+      type="email"
+      matInput
+      placeholder="your email"
+      formControlName="email">
+    <mat-error *ngIf="email?.touched  " class="validation">
+      <small *ngIf="email?.errors?.['required']">Field email should not be empty</small>
+      <small *ngIf="email?.errors?.['email']">Email should be valid</small>
+    </mat-error>
+  </mat-form-field>
+  <mat-form-field appearance="fill">
+    <input
+      type="password"
+      matInput
+      placeholder="your password"
+      formControlName="password">
+    <mat-error *ngIf="password?.touched && password?.invalid" class="validation">
+      <small *ngIf="password?.errors?.['required']">Field should not be empty</small>
+      <small *ngIf="password?.errors?.['minlength']">
+        Password should be not shorter than {{password?.errors?.['minlength'].requiredLength}} symbols.
+        Now it is {{password?.errors?.['minlength'].actualLength}}
+      </small>
+    </mat-error>
+  </mat-form-field>
+  <button
+    type="submit"
+    mat-raised-button
+    color="primary"
+    [disabled]="loginForm.invalid"
+    [routerLink]="['/']">Login</button>
+</form>
