+++ src/app/auth/login/login.component.ts
@@ -1,0 +1,48 @@
+import { Component, OnInit } from '@angular/core';
+import { FormControl, FormGroup, Validators } from "@angular/forms";
+import { User } from '../models/user';
+
+@Component({
+  selector: 'app-login',
+  templateUrl: './login.component.html',
+  styleUrls: ['./login.component.scss']
+})
+export class LoginComponent implements OnInit {
+  loginForm: FormGroup;
+  user!: User;
+  constructor() {
+    this.loginForm = new FormGroup({
+      email : new FormControl('', [
+        Validators.required,
+        Validators.email
+      ]),
+      password : new FormControl('', [
+        Validators.required,
+        Validators.minLength(8)])
+    })
+  }
+
+  ngOnInit(): void {
+
+  }
+
+  get email() {
+    return this.loginForm.get('email')
+  }
+
+  get password() {
+    return this.loginForm.get('password')
+  }
+
+  submit() {
+    if (this.loginForm.invalid) {
+      return;
+    }
+
+    this.user = {
+      email: this.loginForm.value.email,
+      password: this.loginForm.value.password
+    }
+  }
+
+}
