+++ src/app/auth/models/user.ts
@@ -1,0 +1,4 @@
+export interface User {
+    email: string;
+    password: string;
+}
