+++ src/app/header/header.component.ts
@@ -1,0 +1,25 @@
+import { Component, OnInit } from '@angular/core';
+import { Router } from "@angular/router";
+
+@Component({
+  selector: 'app-header',
+  templateUrl: './header.component.html',
+  styleUrls: ['./header.component.scss']
+})
+export class HeaderComponent implements OnInit {
+
+  constructor(private _router: Router) { }
+
+  ngOnInit(): void {
+  }
+
+  onLogin(event: Event) {
+    console.log('login');
+  }
+
+  onLogout(event: Event) {
+    event.preventDefault();
+    this._router.navigate(['/login']);
+  }
+
+}
