+++ src/app/posts/models/post.ts
@@ -1,0 +1,7 @@
+export interface Post {
+  id?: string;
+  date: Date;
+  title: string;
+  author: string;
+  content: string;
+}
