+++ src/app/posts/post.service.ts
@@ -1,0 +1,13 @@
+import { Post } from "./models/post";
+
+export class PostService {
+  private _post!: Post
+
+  set post(selectedPost: Post) {
+    this._post = selectedPost;
+  }
+
+  get post() : Post {
+    return this._post;
+  }
+}
