+++ src/app/posts/post/post.component.scss
