+++ src/app/posts/posts.component.scss
