--- src/app/shared/components/main-layout/header/header.component.html
@@ -1,9 +1,0 @@
-<mat-toolbar color="primary">
-  <div>Blog</div>
-  <div fxFlex fxLayout fxLayoutAlign="flex-end">
-    <ul fxLayout class="navigation-items">
-      <li><a [routerLink]="['/admin', 'login']">Login as administrator</a></li>
-    </ul>
-  </div>
-</mat-toolbar>
-
