--- src/app/shared/components/main-layout/header/header.component.scss
@@ -1,18 +1,0 @@
-mat-toolbar {
-  position: sticky;
-  top: 0;
-  z-index: 1000;
-}
-a {
-  text-decoration: none;
-}
-
-.navigation-items {
-  list-style: none;
-  padding: 0;
-  margin: 0;
-
-  a {
-    color: rgba(255, 255, 255, 0.7);
-  }
-}
