--- src/app/shared/components/main-layout/main-layout.component.html
@@ -1,6 +1,0 @@
-<app-header></app-header>
-<div class="container">
-  <router-outlet></router-outlet>
-</div>
-
-
