--- src/app/shared/components/post/post.component.ts
@@ -1,22 +1,0 @@
-import { Component, Input, OnInit } from '@angular/core';
-import { Post } from "../../interfaces";
-import { PostService } from "../../services/post.service";
-
-@Component({
-  selector: 'app-post',
-  templateUrl: './post.component.html',
-  styleUrls: ['./post.component.scss']
-})
-export class PostComponent implements OnInit {
-  @Input() public post!: Post;
-  @Input() public openButtonIsVisible!: boolean;
-  constructor(private _postService: PostService) { }
-
-  ngOnInit(): void {
-  }
-
-  onOpenPost(selectedPost: Post) {
-    this._postService.post = selectedPost;
-  }
-
-}
