--- src/app/shared/shared.module.ts
@@ -1,46 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { FlexLayoutModule } from "@angular/flex-layout";
-import { MatToolbarModule } from "@angular/material/toolbar";
-import { MatIconModule } from "@angular/material/icon";
-import { MatCardModule } from "@angular/material/card";
-import { MatButtonModule } from "@angular/material/button";
-import { FormsModule, ReactiveFormsModule } from "@angular/forms";
-import { MatFormFieldModule } from "@angular/material/form-field";
-import { MatInputModule } from "@angular/material/input";
-import { PostService } from "./services/post.service";
-import { PostComponent } from "./components/post/post.component";
-import { RouterModule } from "@angular/router";
-
-@NgModule({
-  declarations: [
-    PostComponent
-  ],
-  imports: [
-    CommonModule,
-    FlexLayoutModule,
-    MatToolbarModule,
-    MatIconModule,
-    MatCardModule,
-    MatButtonModule,
-    FormsModule,
-    ReactiveFormsModule,
-    MatFormFieldModule,
-    MatInputModule,
-    RouterModule
-  ],
-  exports: [
-    FlexLayoutModule,
-    MatToolbarModule,
-    MatIconModule,
-    MatCardModule,
-    MatButtonModule,
-    FormsModule,
-    ReactiveFormsModule,
-    MatFormFieldModule,
-    MatInputModule,
-    PostComponent
-  ],
-  providers: [ PostService ]
-})
-export class SharedModule { }
